+++ packages/wallet-ui/src/components/ui/organism/AccountSwitchModal/AccountsHeader/index.ts
@@ -1,0 +1,1 @@
+export { AccountsHeaderView as AccountsHeader } from './AccountsHeader.view';
