+++ packages/wallet-ui/src/components/ui/organism/AccountSwitchModal/HiddenAccountsList/index.ts
@@ -1,0 +1,1 @@
+export { HiddenAccountsListView as HiddenAccountsList } from './HiddenAccountsList.view';
