+++ packages/wallet-ui/src/components/ui/organism/AccountSwitchModal/VisibleAccountsList/index.ts
@@ -1,0 +1,1 @@
+export { VisibleAccountsListView as VisibleAccountsList } from './VisibleAccountsList.view';
